[26.02.18] : - ended work on UserRequestManagerTest

+++ Vocabulary_UI/src/main/java/com/sschudakov/request/UserRequestManager.java
@@ -126,7 +126,7 @@ public class UserRequestManager {
     //create request
     public void createCollection(String name) throws SQLException {
         if (collectionExists(name)) {
-            throw new IllegalAccessError("Collection with name " + name + " already exists");
+            throw new IllegalArgumentException("Collection with name " + name + " already exists");
         }
         this.wordCollectionService.create(name);
     }
@@ -134,11 +134,19 @@ public class UserRequestManager {
 
     //deleteByName request
     public void deleteWord(String word, String language) throws SQLException {
-        this.wordService.delete(wordService.findByValueAndLanguage(word, language).getWordID());
+        if (wordExists(word, language)) {
+            this.wordService.delete(wordService.findByValueAndLanguage(word, language).getWordID());
+        } else {
+            throw new IllegalArgumentException("There is no word " + word + " in " + language);
+        }
     }
 
     public void deleteCollection(String name) throws SQLException {
-        this.wordCollectionService.deleteByName(name);
+        if (collectionExists(name)) {
+            this.wordCollectionService.deleteByName(name);
+        } else {
+            throw new IllegalArgumentException("There is no collection with the name " + name);
+        }
     }
 
 
@@ -153,49 +161,44 @@ public class UserRequestManager {
         }
     }
 
-    /*public void addMeanings(String word, String language, Map<String, Collection<String>> meanings) throws SQLException {
-        Word persistedWord = this.wordService.findByValueAndLanguage(word, language);
-        for (Map.Entry<String, Collection<String>> stringCollectionEntry : meanings.entrySet()) {
-            for (String s : stringCollectionEntry.getValue()) {
-                addMeaning(persistedWord, s, stringCollectionEntry.getKey());
-            }
-        }
-    }*/
 
     public void addMeaning(String word, String wordLanguage, String meaning, String meaningsLanguage) throws SQLException {
-        throw new UnsupportedOperationException();
-    }
-    /*public void removeMeanings(String word, String language, Map<String, Collection<String>> meanings) throws SQLException {
-        for (Map.Entry<String, Collection<String>> stringCollectionEntry : meanings.entrySet()) {
-            for (String s : stringCollectionEntry.getValue()) {
-                this.wmrService.deleteByName(word, language, s, stringCollectionEntry.getKey());
-            }
+        Word foundWord = this.wordService.findByValueAndLanguage(word, wordLanguage);
+        Word foundMeaning = this.wordService.findByValueAndLanguage(meaning, meaningsLanguage);
+        if (!wordMeaningRelationshipExists(word, wordLanguage, meaning, meaningsLanguage)) {
+            this.wmrService.create(foundWord, foundMeaning);
+        } else {
+            throw new IllegalArgumentException("Word " + word + " already has meaning " + meaning);
         }
-    }*/
-    public void removeMeaning(String word, String wordLanguage, String meaning, String meaningsLanguage) throws SQLException{
-        throw new UnsupportedOperationException();
-    }
-    /*public void putInCollections(String word, String lanugage, Collection<String> collections) throws SQLException {
-        Word foundWord = this.wordService.findByValueAndLanguage(word, lanugage);
-        for (String collection : collections) {
-            this.wcrService.create(foundWord, this.wordCollectionService.findByName(collection));
-        }
-    }*/
-    public void putInCollection(String word, String wordLanguage, String collection) throws SQLException {
-        this.wcrService.create(
-                this.wordService.findByValueAndLanguage(word,wordLanguage),
-                this.wordCollectionService.findByName(collection)
-        );
-    }
-    /*public void removeFromCollections(String word, String language, Collection<String> collections) throws SQLException {
-        for (String collection : collections) {
-            removeFromCollection(word, language, collection);
-        }
-    }*/
-    public void removeFromCollection(String word, String language, String collectionName) throws SQLException {
-        this.wcrService.delete(word, language, collectionName);
     }
 
+    public void removeMeaning(String word, String wordLanguage, String meaning, String meaningsLanguage) throws SQLException {
+        if (wordMeaningRelationshipExists(word, wordLanguage, meaning, meaningsLanguage)) {
+            this.wmrService.delete(word, wordLanguage, meaning, meaningsLanguage);
+        } else {
+            throw new IllegalArgumentException("word " + word + " has no meaning " + meaning);
+        }
+    }
+
+
+    public void putInCollection(String word, String wordLanguage, String collection) throws SQLException {
+        if (!wordCollectionsRelationshipExists(word, wordLanguage, collection)) {
+            this.wcrService.create(
+                    this.wordService.findByValueAndLanguage(word, wordLanguage),
+                    this.wordCollectionService.findByName(collection)
+            );
+        } else {
+            throw new IllegalArgumentException("word " + word + " is already in " + collection + " collection");
+        }
+    }
+
+    public void removeFromCollection(String word, String wordLanguage, String collection) throws SQLException {
+        if (wordCollectionsRelationshipExists(word, wordLanguage, collection)) {
+            this.wcrService.delete(word, wordLanguage, collection);
+        } else {
+            throw new IllegalArgumentException("there is no word " + word + " in " + collection + " collection");
+        }
+    }
 
 
     //check requests
@@ -207,14 +210,11 @@ public class UserRequestManager {
         return wordCollectionService.findByName(collection) != null;
     }
 
-    private boolean wordMeaningRelationshipExists(Word word, Word meaning) {
-        throw new UnsupportedOperationException();
-        /*return this.wmrService.findByWordAndMeaningIds(word.getWordID(), meaning.getWordID()) != null;*/
+    private boolean wordMeaningRelationshipExists(String word, String language, String meaning, String meaningLanguage) throws SQLException {
+        return this.wmrService.findByWordAndMeaning(word, language, meaning, meaningLanguage).size() == 2;
     }
 
-    private boolean wordCollectionsRelationshipExists(Word word, WordCollection collection) {
-        throw new UnsupportedOperationException();
-        /*return this.wcrService.findByWordAndCollectionIds(word.getWordID(),
-                collection.getCollectionID()) != null;*/
+    private boolean wordCollectionsRelationshipExists(String word, String language, String collection) throws SQLException {
+        return this.wcrService.findByWordAndCollection(word, language, collection) != null;
     }
 }
